[lldb] Fix build failure introduced by 3d4cadfb26437bd686ca8177f5454a366fed59eb
This patch updates the `ScriptedProcess::GetGenericInteger` return type
to `llvm::Optional<unsigned long long>` to match implementation.

Differential Revision: https://reviews.llvm.org/D105788

+++ lldb/source/Plugins/ScriptInterpreter/Python/ScriptedProcessPythonInterface.h
@@ -51,7 +51,8 @@ public:
   bool IsAlive() override;
 
 protected:
-  llvm::Optional<uint64_t> GetGenericInteger(llvm::StringRef method_name);
+  llvm::Optional<unsigned long long>
+  GetGenericInteger(llvm::StringRef method_name);
   Status GetStatusFromMethod(llvm::StringRef method_name);
 
 private:
